Update optional and variant to disambiguate in-plate_t usage.
- Had to centralize the in_place support for windows builds.
