+++ vital/internal/variant/variant.hpp
@@ -206,30 +206,10 @@ namespace mpark = kwiver::vital;
 
 #include "config.hpp"
 #include "lib.hpp"
+#include "../in_place.hpp"
 
-// namespace mpark {
 namespace kwiver {
 namespace vital{
-
-
-// in-place support
-  struct in_place_t { explicit in_place_t() = default; };
-
-  template <std::size_t I>
-  struct in_place_index_t { explicit in_place_index_t() = default; };
-
-  template <typename T>
-  struct in_place_type_t { explicit in_place_type_t() = default; };
-
-#ifdef MPARK_VARIABLE_TEMPLATES
-  constexpr in_place_t in_place{};
-
-  template <std::size_t I> constexpr in_place_index_t<I> in_place_index{};
-
-  template <typename T> constexpr in_place_type_t<T> in_place_type{};
-#endif
-
-
 
 #ifdef MPARK_RETURN_TYPE_DEDUCTION
